Experimental Build (v0.1.1-Ealpha) Snapshot 1s2a
Reaction roles are in progress of being made.

Added command to assign emojis, added commnand to assign reactions and to send the reaction role message. For some reason the process of giving out roles is a bit bugged and it works on some servers and on some doesn't.

Bug removal process in progress. Development also in progress.